Refactoring of some of the classes
